Add LoadSongs and loads of stuff, see description
GameMaster.LoadSongs() will parse the Songs and add them to the
GameMaster at runtime.
Songs are now located in "Resources/Songs" and should be copied to
output directory (see "Adding songs in Visual Studio.txt").
Songs need a title, author, year, difficulty and instrument list to be
instantiated.
Instruments are now an Enum.